--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -130,16 +130,69 @@ jobs:
       - name: Test --features strict
         run: timeout 15m cargo test -p interledger-btp --features strict
 
-  test-md:
+
+  settlement-engines:
+    # This job builds the ethereum-settlement-engine required by `test-md`.
+    # A separate job to enable dependency caching.
     runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout sources
+        uses: actions/checkout@v2
+        with:
+          repository: interledger-rs/settlement-engines
+          ref: master
+
+      - name: Install rust toolchain
+        uses: hecrj/setup-rust-action@v1.3.4
+        with:
+          rust-version: stable
+
+      - name: Cache
+        uses: Swatinem/rust-cache@v1
+
+      - name: Build
+        run: cargo build --bin ilp-settlement-ethereum
+
+      - name: Upload settlement-engine artifact
+        uses: actions/upload-artifact@v2
+        with:
+          name: settlement-engine
+          path: target/debug/ilp-settlement-ethereum
+
+  test-md:
+    env:
+      RUST_LOG: "interledger=trace"
+      RUST_BACKTRACE: "full"
+
+    needs: settlement-engines
+
+    runs-on: ubuntu-latest
+
     steps:
       - name: Checkout sources
         uses: actions/checkout@v2
 
-      - name: Install node
-        uses: actions/setup-node@v2
+      - name: Install rust toolchain
+        uses: hecrj/setup-rust-action@v1.3.4
         with:
-          node-version: 'v12.18.4'
+          rust-version: stable
+
+      - name: Cache
+        uses: Swatinem/rust-cache@v1
+
+      - name: Build binaries
+        run: |
+          mkdir -p ~/.interledger/bin
+          cargo build --bin ilp-node --bin ilp-cli
+          mv target/debug/ilp-node ~/.interledger/bin
+          mv target/debug/ilp-cli ~/.interledger/bin
+
+      - name: Download settlement engine artifact
+        uses: actions/download-artifact@v2
+        with:
+          name: settlement-engine
+          path: ~/.interledger/bin
 
       - name: Install dependencies
         run: |
@@ -150,8 +203,7 @@ jobs:
           npm install -g ganache-cli ilp-settlement-xrp conventional-changelog-cli
 
       - name: Test
-        run: |
-          scripts/run-md-test.sh '^.*$' 1
+        run: scripts/run-md-test.sh '^.*$' 2
 
       - name: 'Store artifacts'
         if: ${{ failure() }}