This PR adds dependency caching to the CI workflow, dramatically speeding up builds (after the first one). Note: does not include the docker workflow -- may be of interest too (?) but I'm not sure how it could work.

Also adds `workflow_dispatch` trigger, which will let us manually trigger the workflow. Mainly useful for testing on forks without having to fiddle with the workflow file.

Uses the [rust-cache action](https://github.com/Swatinem/rust-cache). We could try manually figure out which parts of `/target` to cache, but I found it difficult and prone to error. This action seems to do the right thing and ends up caching about 2 GB total with the way I have it setup (multiple jobs, with independent caches).

It may be possible to further optimize the cache by sharing amongst the jobs, but I'm not quite certain how.

What I found surprising is that the `clippy` check is so slow -- it seems to be checking *all* depedencies. Might be worth investigating. I did a small search which turned up the `--no-deps` option but it didn't change anything for me. This is only for a non-cached run -- since the cache will include the dependency clippy checks for the next run.

Comparison between runs:
| Run  | Duration | Cumulative time |
| --- | --- | --- |
| [before](https://github.com/interledger-rs/interledger-rs/actions/runs/1093773784) | 17m 39s | 33m 37s |
| [no cache](https://github.com/Mirko-von-Leipzig/interledger-rs/actions/runs/1104556424) | 12m 51s | 30m 1s |
| [cache](https://github.com/Mirko-von-Leipzig/interledger-rs/actions/runs/1104557123) | 7m 54s | 15m 47s |

Cache size per job:
| Job | Cache |
| --- | --- |
| Interledger | ~440 MB |
| Interledger-packet | ~166 MB |
| Interledger-stream | ~251 MB |
| Interledger-btp | ~294 MB |
| Settlement-engines | ~364 MB |
| test-md | ~335 MB |
| **Total** | **~1.85 GB** |

Some restrictions on [github action caching](https://docs.github.com/en/actions/guides/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy):
- maximum size of 5GB *total* per repo - replaces oldest cache automatically on overflow
- gets cleared automatically after 7 days of inactivity