@koivunej 2021-08-06
> What I found surprising is that the clippy check is so slow -- it seems to be checking all depedencies.

I think it wont benefit much from the previous `cargo build` above it because it'll do the `--all-targets` dance but it might help the `cargo test` which comes after it.