@koivunej 2021-08-06
What's still missing from the CI workflow which I suddenly remembered is checking the fuzz targets. I don't think it can be done without `cargo +nightly fuzz build`, which I think `serde/json` was also using. This would also add nightly compiler installation together with `cargo install cargo-fuzz` which costs around 3min of user time for me which would be I guess 90s with 2 cpus.

Could just create a new issue out of this.