@koivunej 2021-08-06
Latest run in 7min 7s, very nice.