@koivunej 2021-08-06
@Mirko-von-Leipzig could you restore a workflow called "build" to be a dummy `Always succeed` because this repository is configured with both of the named workflows being required for merge? :)